requirements.txt
aniso8601==9.0.1
attrs==22.2.0
autopep8==2.0.2
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.2
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
Werkzeug==2.2.3
